DDCNL-1725 | AddressSubmissionController created
