+++ app/views/personaldetails/reviewChanges.scala.html
@@ -91,7 +91,7 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
           }
       </dl>
 
-      @form(controllers.routes.AddressController.submitChanges(typ)) {
+      @form(controllers.address.routes.AddressSubmissionController.onSubmit(typ)) {
         <button type="submit">@messages("label.confirm_and_save")</button>
       }
     </div>
