+++ test/controllers/AddressControllerSpec.scala
@@ -157,481 +157,6 @@ class AddressControllerSpec extends BaseSpec with MockitoSugar {
     })
   }
 
-  "Calling AddressController.reviewChanges" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-    }
-
-    "return 200 if both SubmittedAddressDto and SubmittedStartDateDto are present in keystore for non-postal" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "return 200 if only SubmittedAddressDto is present in keystore for postal" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "postalSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )))
-
-      val result = controller.reviewChanges(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "redirect back to start of journey if SubmittedAddressDto is missing from keystore for non-postal" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )))
-
-      val result = controller.reviewChanges(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "redirect back to start of journey if SubmittedAddressDto is missing from keystore for postal" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map()
-        )
-      )
-
-      val result = controller.reviewChanges(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "display no message relating to the date the address started when the primary address has not changed" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(PrimaryAddrType)(FakeRequest())
-      //  implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) shouldNot include(controller.messagesApi("label.when_this_became_your_main_home"))
-    }
-
-    "display no message relating to the date the address started when the primary address has not changed when the postcode is in lower case" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodifiedLowerCase)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(PrimaryAddrType)(FakeRequest())
-      implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) shouldNot include(Messages("label.when_this_became_your_main_home"))
-    }
-
-    "display no message relating to the date the address started when the primary address has not changed when the postcode entered has no space" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto" -> Json.toJson(
-              asAddressDto(fakeStreetTupleListAddressForUnmodifiedNoSpaceInPostcode)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(PrimaryAddrType)(FakeRequest())
-      implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) shouldNot include(Messages("label.when_this_became_your_main_home"))
-    }
-
-    "display a message relating to the date the address started when the primary address has changed" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForModifiedPostcode)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(PrimaryAddrType)(FakeRequest())
-      implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) should include(Messages("label.when_this_became_your_main_home"))
-    }
-
-    "display the appropriate label for address when the sole address has changed" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForModifiedPostcode)),
-            "soleSubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(SoleAddrType)(FakeRequest())
-      implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) should include(Messages("label.your_new_address"))
-      contentAsString(result) should include(Messages("label.when_you_started_living_here"))
-    }
-
-    "display the appropriate label for address when the sole address has not changed" in new LocalSetup {
-      lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "soleSubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      val result = controller.reviewChanges(SoleAddrType)(FakeRequest())
-      implicit val messages: Messages = Messages.Implicits.applicationMessages
-
-      contentAsString(result) should include(Messages("label.your_address"))
-      contentAsString(result) shouldNot include(Messages("label.when_you_started_living_here"))
-    }
-  }
-
-  "Calling AddressController.submitChanges" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-
-      def comparatorDataEvent(
-        dataEvent: DataEvent,
-        auditType: String,
-        uprn: Option[String],
-        includeOriginals: Boolean,
-        submittedLine1: Option[String] = Some("1 Fake Street"),
-        addressType: Option[String] = Some("Residential")) = DataEvent(
-        "pertax-frontend",
-        auditType,
-        dataEvent.eventId,
-        Map("path" -> "/test", "transactionName" -> "change_of_address"),
-        Map(
-          "nino"              -> Some(Fixtures.fakeNino.nino),
-          "etag"              -> Some("115"),
-          "submittedLine1"    -> submittedLine1,
-          "submittedLine2"    -> Some("Fake Town"),
-          "submittedLine3"    -> Some("Fake City"),
-          "submittedLine4"    -> Some("Fake Region"),
-          "submittedPostcode" -> Some("AA1 1AA"),
-          "submittedCountry"  -> None,
-          "addressType"       -> addressType,
-          "submittedUPRN"     -> uprn,
-          "originalLine1"     -> Some("1 Fake Street").filter(x => includeOriginals),
-          "originalLine2"     -> Some("Fake Town").filter(x => includeOriginals),
-          "originalLine3"     -> Some("Fake City").filter(x => includeOriginals),
-          "originalLine4"     -> Some("Fake Region").filter(x => includeOriginals),
-          "originalPostcode"  -> Some("AA1 1AA").filter(x => includeOriginals),
-          "originalCountry"   -> Some("Country(UK,United Kingdom)").filter(x => includeOriginals),
-          "originalUPRN"      -> uprn.filter(x => includeOriginals)
-        ).map(t => t._2.map((t._1, _))).flatten.toMap,
-        dataEvent.generatedAt
-      )
-    }
-
-    "redirect to start of journey if primarySubmittedStartDateDto is missing from the cache, and the journey type is PrimaryAddrType" in new LocalSetup {
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]])
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-
-      verify(mockAuditConnector, times(0)).sendEvent(any())(any(), any())
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(controller.editAddressLockRepository, times(0)).insert(meq(nino.withoutSuffix), meq(SoleAddrType))
-
-    }
-
-    "redirect to start of journey if soleSubmittedStartDateDto is missing from the cache, and the journey type is SoleAddrType" in new LocalSetup {
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "soleSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(SoleAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-
-      verify(mockAuditConnector, times(0)).sendEvent(any())(any(), any())
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "render the thank-you page if postalSubmittedStartDateDto is not in the cache, and the journey type is PostalAddrType" in new LocalSetup {
-      override lazy val fakeAddress =
-        buildFakeAddress.copy(`type` = Some("Correspondence"), startDate = Some(LocalDate.now))
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "postalSubmittedAddressDto" -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-    }
-
-    "redirect to start of journey if primarySubmittedAddressDto is missing from the cache" in new LocalSetup {
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-
-      verify(mockAuditConnector, times(0)).sendEvent(any())(any(), any())
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "render the thank-you page and log a postcodeAddressSubmitted audit event upon successful submission of an unmodified address" in new LocalSetup {
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySelectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord),
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(
-        dataEvent,
-        "postcodeAddressSubmitted",
-        Some("GB101"),
-        false)
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-    }
-
-    "render the thank you page and log a postcodeAddressSubmitted audit event upon successful submission of an unmodified address, this time using postal type and having no postalSubmittedStartDateDto in the cache " in new LocalSetup {
-      override lazy val fakeAddress =
-        buildFakeAddress.copy(`type` = Some("Correspondence"), startDate = Some(LocalDate.now))
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "postalSelectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord),
-            "postalSubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForUnmodified))
-          )
-        ))
-
-      val requestWithForm: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", "/test")
-          .withFormUrlEncodedBody(fakeStreetTupleListAddressForUnmodified: _*)
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              request = requestWithForm
-                .asInstanceOf[Request[A]])
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(
-        dataEvent,
-        "postcodeAddressSubmitted",
-        Some("GB101"),
-        false,
-        addressType = Some("Correspondence"))
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-    }
-
-    "render the thank you page and log a manualAddressSubmitted audit event upon successful submission of a manually entered address" in new LocalSetup {
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForManualyEntered)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(dataEvent, "manualAddressSubmitted", None, false)
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-    }
-
-    "render the thank you page and log a postcodeAddressModifiedSubmitted audit event upon successful of a modified address" in new LocalSetup {
-      override lazy val fakeAddress = buildFakeAddress.copy(line1 = Some("11 Fake Street"))
-      override lazy val sessionCacheResponse = Some(
-        CacheMap(
-          "id",
-          Map(
-            "primarySelectedAddressRecord" -> Json.toJson(fakeStreetPafAddressRecord),
-            "primarySubmittedAddressDto"   -> Json.toJson(asAddressDto(fakeStreetTupleListAddressForModified)),
-            "primarySubmittedStartDateDto" -> Json.toJson(DateDto.build(15, 3, 2015))
-          )
-        ))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitChanges(PrimaryAddrType)(FakeRequest())
-
-      status(result) shouldBe OK
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(
-        dataEvent,
-        "postcodeAddressModifiedSubmitted",
-        Some("GB101"),
-        true,
-        Some("11 Fake Street"))
-      verify(controller.sessionCache, times(1)).fetch()(any(), any())
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-    }
-  }
-
   "Calling AddressController.showAddressAlreadyUpdated" should {
 
     trait LocalSetup extends WithAddressControllerSpecSetup {
